+++ R/average_loss.R
@@ -20,14 +20,15 @@
 #'   a (n x K) matrix of probabilities (with row-sums 1).
 #'   The observed values `y` are either passed as (n x K) dummy matrix, 
 #'   or as discrete vector with corresponding levels. 
-#'   The latter case is turned into a dummy matrix via `model.matrix(~ y + 0)`.
-#' - "classification_error": Misclassification error. This is currently the
-#'   only prespecified loss that accepts non-numeric predictions. Both the 
+#'   The latter case is turned into a dummy matrix via 
+#'   `model.matrix(~ as.factor(y) + 0)`.
+#' - "classification_error": Misclassification error. Both the 
 #'   observed values `y` and the predictions can be character/factor. This
 #'   loss function can be used in non-probabilistic classification settings.
-#'   BUT: Probabilistic classification (with "mlogloss") is preferred.
+#'   BUT: Probabilistic classification (with "mlogloss") is clearly preferred in most
+#'   situations.
 #' - A function with signature `f(actual, predicted)`, returning a numeric 
-#'   vector of length n or a matrix with n rows, where n is the number of rows of `X`.
+#'   vector or matrix of the same length as the input.
 #'
 #' @inheritParams hstats
 #' @param y Vector/matrix of the response corresponding to `X`.
@@ -36,7 +37,7 @@
 #'   can be provided that turns observed and predicted values into a numeric vector or 
 #'   matrix of unit losses of the same length as `X`.
 #'   For "mlogloss", the response `y` can either be a dummy matrix or a discrete vector. 
-#'   The latter case is handled via `model.matrix(~ y + 0)`.
+#'   The latter case is handled via `model.matrix(~ as.factor(y) + 0)`.
 #'   For "classification_error", both predictions and responses can be non-numeric.
 #' @param BY Optional grouping vector.
 #' @returns A matrix with one row per group and one column per loss dimension.
@@ -73,9 +74,6 @@ average_loss.default <- function(object, X, y,
       is.vector(BY) || is.factor(BY),
       length(BY) == nrow(X)
     )
-  }
-  if (!is.function(loss) && loss == "mlogloss" && NCOL(y) == 1L) {
-    y <- stats::model.matrix(~y + 0)
   }
   if (!is.function(loss)) {
     loss <- get_loss_fun(loss)